sortMedias switch fixed for title comparison
